logging_tracing

Module prelude

Re-exports

pub use crate::LogMode;

pub use crate::TraceScope;

pub use crate::TracingLibraryBuilder;

Modules

instrument

Attach a span to a std::future::Future.

span

Spans represent periods of time in which a program was executing in a particular context.

Macros

debug

Constructs an event at the debug level.

error

Constructs an event at the error level.

info

Constructs an event at the info level.

span

Constructs a new span.

trace

Constructs an event at the trace level.

warn

Constructs an event at the warn level.

Structs

Level

Describes the level of verbosity of a span or event.

Attribute Macros

instrument

Instruments a function to create and enter a tracing span every time the function is called.